style fixes and some refactoring to pgcr
