+++ src/styles/pgcr.module.css
@@ -90,7 +90,7 @@
 }
 
 .cp-error {
-    min-width: 50%;
+    min-width: 40%;
     max-width: 70%;
 }
 
@@ -152,15 +152,19 @@
     width: 100%;
     height: 100%;
 
-    opacity: 0.95;
+    background-color: var(--modal-background);
+    opacity: 0.9;
     overflow: hidden;
 }
 
 .emblem img {
+    filter: brightness(0.9) blur(1px);
     max-height: 100%;
 }
 
 .member-card {
+    position: absolute;
+    top: 0;
     z-index: 1;
     width: 100%;
     height: 100%;
@@ -335,7 +339,7 @@
 
 .class-button-container {
     height: auto;
-    flex: 1;
+    flex: 0;
 
     margin: 2px;
     padding: 0;
@@ -345,11 +349,14 @@
     flex-wrap: nowrap;
 
     justify-content: space-evenly;
+    align-items: center;
     gap: 0.8em;
 }
 
 .class-button {
     padding: 0.5em;
+    width: 80px;
+    height: 80px;
 
     background-color: rgba(151, 151, 151, 0.274);
 
@@ -359,14 +366,13 @@
 }
 
 .class-button > img {
-    max-width: 70px;
-    aspect-ratio: 1;
-    margin: auto;
+    width: 100%;
+    height: 100%;
 }
 
 .member-profile-button {
     height: 85px;
-    flex: 1;
+    flex-grow: 0;
     background-color: rgba(55, 85, 79, 0.287);
 
     display: flex;
@@ -375,10 +381,11 @@
 }
 
 .member-profile-link {
-    margin-right: 0.7em;
     display: flex;
     align-items: center;
     justify-content: center;
+    gap: 0.5em;
+
     text-decoration: none;
 }
 
@@ -387,14 +394,16 @@
     max-height: 30px;
     max-width: 30px;
     margin: 10px;
-    margin-right: 15px;
 }
 
 .member-profile-link > span {
     flex: 1;
+    margin-right: 0.7em;
+
+    display: inline-flex;
+
     color: white;
     font-size: 1.15rem;
-    display: inline-flex;
 }
 
 .summary-stats {
@@ -412,7 +421,7 @@
 }
 
 .summary-stat {
-    min-height: 57px;
+    min-height: 50px;
     padding: 10px 15px;
     background-color: rgba(78, 9, 9, 0.35);
     display: flex;
@@ -503,9 +512,22 @@
     .raid-name > span {
         font-size: x-large;
     }
-
+    .cp-error {
+        max-width: 40%;
+    }
     .duration > span {
         font-size: large;
+    }
+    .selected-entry-card,
+    .member-profile-button {
+        height: 69px;
+    }
+    .class-button {
+        width: 60px;
+        height: 60px;
+    }
+    .member-profile-link > span {
+        display: none;
     }
     .members {
         grid-template-columns: 1fr;
